migrate home page
