--- apps/web/src/pages/Home/HomePage/components/About/index.ts
@@ -1,3 +1,0 @@
-import About from "./About";
-
-export default About;
