--- apps/web/src/pages/Home/HomePage/components/Featured/index.ts
@@ -1,3 +1,0 @@
-import Featured from "./Featured";
-
-export default Featured;
