--- apps/web/src/pages/Home/HomePage/components/Hero/index.ts
@@ -1,3 +1,0 @@
-import Hero from "./Hero";
-
-export default Hero;
