--- apps/web/src/pages/Home/HomePage/components/Opportunities/index.ts
@@ -1,3 +1,0 @@
-import Opportunities from "./Opportunities";
-
-export default Opportunities;
