--- apps/web/src/pages/Home/HomePage/components/index.ts
@@ -1,7 +1,0 @@
-import About from "./About";
-import Featured from "./Featured";
-import Hero from "./Hero";
-import Opportunities from "./Opportunities";
-import Profile from "./Profile";
-
-export { About, Featured, Hero, Opportunities, Profile };
